workers/app/main.py
from celery import Celery
from dotenv import load_dotenv
import os

load_dotenv(os.path.join(os.path.dirname(__file__), '..', '..', '.env'))

celery_app = Celery(
    "workers",
    broker="redis://localhost:6379/0",
    backend="redis://localhost:6379/0",
    include=["app.tasks"]
)
celery_app.conf.update(task_track_started=True)