workers/requirements.txt
celery
redis
python-dotenv
google-generativeai